imported sidepanel, added outlets,moved patientProfile.jsx code into patientHomePage.jsx
